Make system subset selection generic
Selection of pairs is really just selection of a 2-subset. The subset
size is now a parameter. This also alleviates to an extent the need
for a specialised system difference iterator: it's just a special case
of the subset selection. Since it's not clear that all users need it,
it's left out for now.